# lib: optimize common prefix computation of two hex strings